Refactor: move toast utility functions to it on js file.
