+++ resources/js/application-page.js
@@ -1,9 +1,9 @@
+import { formatNumberToCurrency } from './Utilities/utilFunctions';
 import {
     showToastFeedback,
     showProcessToast,
     hideProcessToast,
-    formatNumberToCurrency,
-} from './Utilities/utilFunctions';
+} from './Utilities/feedback-toast';
 import { customFormatNumericInput } from './Utilities/input-utils';
 import { FormDraftHandler } from './Utilities/FormDraftHandler';
 import {
