+++ resources/js/cooperator/coop-page.js
@@ -2,15 +2,17 @@ import '../echo';
 import { customFormatNumericInput } from '../Utilities/input-utils';
 import createConfirmationModal from '../Utilities/confirmation-modal';
 import {
-    showToastFeedback,
     customDateFormatter,
     formatNumberToCurrency,
     parseFormattedNumberToFloat,
-    showProcessToast,
-    hideProcessToast,
     closeModal,
     serializeFormData,
 } from '../Utilities/utilFunctions';
+import {
+    showProcessToast,
+    hideProcessToast,
+    showToastFeedback,
+} from '../Utilities/feedback-toast';
 import { FormDraftHandler } from '../Utilities/FormDraftHandler';
 import QUARTERLY_REPORTING_FORM_CONFIG from '../Form_Config/QUARTERLY_REPORTING_CONFIG';
 import {
